# docs/swagger.yaml
basePath: /api/v1
definitions:
  handlers.AccountTransferRequest:
    properties:
      amount:
        type: number
      date:
        type: string
      note:
        type: string
      to_account_id:
        type: string
    required:
    - amount
    - date
    - to_account_id
    type: object
...
  handlers.CreateAccountRequest:
    properties:
      color:
        type: string
      description:
        type: string
      goal_amount:
        description: SAVINGS
        type: number
      goal_label:
        description: SAVINGS
        type: string
      income_type:
        description: required for CARD
        type: string
      initial_balance:
        type: number
      name:
        type: string
      type:
        description: CARD|CASH|SAVINGS
        type: string
    required:
    - name
    - type
    type: object
...
  handlers.SpentRequest:
    properties:
      amount:
        type: number
      category_id:
        type: string
      date:
        type: string
      description:
        type: string
    required:
    - amount
    - category_id
    - date
    type: object
...
  handlers.TopUpRequest:
    properties:
      amount:
        type: number
      date:
        type: string
      description:
        type: string
      source:
        type: string
    required:
    - amount
    - date
    - source
    type: object
  handlers.UpdateAccountRequest:
    properties:
      color:
        type: string
      description:
        type: string
      goal_amount:
        description: SAVINGS
        type: number
      goal_label:
        description: SAVINGS
        type: string
      income_type:
        description: CARD only
        type: string
      name:
        type: string
    type: object
...
  models.ErrorResponse:
    properties:
      error:
        type: string
      success:
        type: boolean
    type: object
...
  models.Response:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
host: localhost:8081
info:
  contact:
    email: support@financeapp.com
    name: API Support
    url: https://github.com/yourusername/finance-tracking-app
  description: A comprehensive personal finance tracking application API with budget
    management, expense tracking, and analytics
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Finance Tracking API
  version: "1.0"
paths:
  /accounts:
    post:
      consumes:
      - application/json
      parameters:
      - description: Account data
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateAccountRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Create a new account
      tags:
      - accounts
  /accounts/{id}:
    get:
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
      summary: Get account by ID
      tags:
      - accounts
    patch:
      consumes:
      - application/json
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      - description: Account update data
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateAccountRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
      summary: Update account details
      tags:
      - accounts
  /accounts/{id}/archive:
    post:
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
      summary: Archive an account
      tags:
      - accounts
  /accounts/{id}/spent:
    post:
      consumes:
      - application/json
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      - description: Expense data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.SpentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Response'
      summary: Record an expense from an account
      tags:
      - accounts
  /accounts/{id}/summary:
    get:
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      - description: Month (1-12)
        in: query
        name: month
        type: integer
      - description: Year
        in: query
        name: year
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
      summary: Get account balance summary for a specific month
      tags:
      - accounts
  /accounts/{id}/topup:
    post:
      consumes:
      - application/json
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      - description: Top-up data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.TopUpRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Response'
      summary: Add funds (income) to an account
      tags:
      - accounts
  /accounts/{id}/transfer:
    post:
      consumes:
      - application/json
      parameters:
      - description: Source Account ID
        in: path
        name: id
        required: true
        type: string
      - description: Transfer data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.AccountTransferRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Response'
      summary: Transfer funds from this account to another
      tags:
      - accounts
  /accounts/{id}/transfers:
    get:
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      - description: Month (1-12)
        in: query
        name: month
        type: integer
      - description: Year
        in: query
        name: year
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
      summary: Get transfer history for a specific account
      tags:
      - accounts
...
schemes:
- http
- https
swagger: "2.0"